Toolbar color state position fix
Now it's based on the scrollview Y position, calling scrollTo in a runnable is a lot more reliable and keeps the image parallax aligned correctly as well since the obersvableScrollView interfaces are already set to immedately update on it's scroll change.

Also cleaned up instance state handing and removed unused onFragmentInteraction interface.